refactor(config): extract types and load modules
Split the remaining ~3.7k-line config/mod.rs into types.rs (public
config structs and their impls) and load.rs (Config::load and target
resolution helpers). mod.rs keeps ConfigError, curated re-exports,
and the nested test module.
