+++ crates/rag-rat-core/src/config/mod.rs
@@ -1,1213 +1,13 @@
-use std::collections::{BTreeMap, BTreeSet};
-use std::fs;
-use std::path::{Path, PathBuf};
-use std::str::FromStr;
-use std::time::Duration;
+use std::path::PathBuf;
 
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-use crate::embedding_models::{
-    Backend, EmbeddingModelSpec, FASTEMBED_MODEL_ID, MODEL2VEC_MODEL_ID, spec,
-};
-use crate::language::{Language, LanguageError};
-
-#[derive(Debug, Clone)]
-pub struct Config {
-    pub root: PathBuf,
-    pub database: PathBuf,
-    pub targets: Vec<ResolvedTarget>,
-    pub llm: LlmConfig,
-    pub watch: WatchConfig,
-    pub log: LogConfig,
-    pub version_check: VersionCheckConfig,
-    pub oracle: OracleConfig,
-    pub search: SearchConfig,
-    pub memory: MemoryConfig,
-    /// Optional `[index] repo_id` override for the consolidated global store — pins the repo's
-    /// identity instead of deriving it from the root-commit hash. `None` = derive. Consumed by
-    /// `crate::repo_identity::resolve_repo_identity`. An EXPLICIT `database` path never depends on
-    /// it; a KEYLESS config's default does only through the identity-existence gate (a pin makes a
-    /// non-git root identity-bearing, so it resolves to the global store instead of per-root).
-    pub repo_id_override: Option<String>,
-    /// Whether the GOVERNING config sets an explicit `[index] database` key. "Governing" is
-    /// main-worktree-anchored (see `main_database_key`): in a linked worktree the MAIN checkout's
-    /// config decides, exactly as it does for `repo_id` — a branch-local config can neither pin
-    /// nor un-pin the repo's database. `rag-rat consolidate` keys its pinned-config refusal off
-    /// this, so the refusal and `database` resolution can never disagree.
-    pub database_key_pinned: bool,
-    /// When `Config::load` re-anchored `[index] root` from a LINKED worktree to the MAIN checkout
-    /// (so all worktrees share one base index), this holds the ORIGINAL linked-worktree root; else
-    /// `None`. The pre-anchor root is otherwise lost after `anchor_root_to_main_worktree`. Used by
-    /// the `index` command to warn that it is indexing the main checkout, not the named worktree
-    /// (#427). Not read from TOML — populated during load.
-    pub source_root_reanchored_from: Option<PathBuf>,
-    /// Opt in to registering an EMPTY index (zero discovered files). Default `false`: the core
-    /// registration path (`rebuild_with_progress`) refuses a first-time-empty registration with
-    /// [`crate::index::EmptyIndexRefused`], so no entry point silently creates one (#427). The CLI
-    /// `index --allow-empty` flag sets this `true`; every other caller (watcher, git-hook
-    /// `maintenance`, MCP, init) leaves it `false`. Not read from TOML — set per invocation.
-    pub allow_empty: bool,
-}
-
-/// Search-ranking knobs (`[search]`). Default OFF so the shipped fuse is byte-identical to today;
-/// opt in per-repo via `rag-rat.toml`.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct SearchConfig {
-    /// Replace the binary git has-history boost with a graded recency+churn magnitude and add the
-    /// generated/test demotion at the wide-pool rerank site (default false — opt in explicitly).
-    /// A/B-swept on the commit-replay eval (`rag-rat eval --replay --rerank`); see
-    /// [`crate::search::lexical::SearchOptions::graded_history`].
-    pub graded_git_rerank: bool,
-}
-
-/// Repo-memory surfacing (`[memory]`) — how memory attachments and memory-query results render.
-/// Default is `summary`: an LLM-compacted summary + verdict marker (dream-generated) instead of the
-/// full body, so a drive-by attachment or a `memory_search` hit is scannable rather than an
-/// 8000-char wall. `full` restores whole bodies everywhere. Either way, `memory show` /
-/// `memory_show` always return the full body — that is the deliberate "expand on request" path.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct MemoryConfig {
-    pub surface: MemorySurface,
-}
-
-/// `[memory] surface` — the memory rendering mode. `Summary` (default) renders the dream-compacted
-/// summary (when one exists for the memory's current body) plus a plain-text verdict marker,
-/// falling back to the mechanical compact header (title-only) when no summary row exists; `Full`
-/// keeps whole bodies. Persisted only in config, so a closed enum with a stable `as_str` round-trip
-/// is enough (no DB column).
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub enum MemorySurface {
-    /// Title + compacted summary + verdict marker, falling back to title-only when no summary row
-    /// exists for the current body. The default — bodies are deferred to `memory show`.
-    #[default]
-    Summary,
-    /// The full body (plus the mechanical compact header where a compact projection is emitted) —
-    /// whole prose everywhere, the pre-summary behavior.
-    Full,
-}
-
-impl MemorySurface {
-    /// The stable config string (matches the toml `surface = "..."` value).
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Full => "full",
-            Self::Summary => "summary",
-        }
-    }
-
-    /// Parse a `surface = "..."` value (case-insensitive). `None` for an unrecognized value — the
-    /// config layer turns that into `ConfigError::UnknownMemorySurface`.
-    fn parse(s: &str) -> Option<Self> {
-        match s.trim().to_ascii_lowercase().as_str() {
-            "full" => Some(Self::Full),
-            "summary" => Some(Self::Summary),
-            _ => None,
-        }
-    }
-}
-
-/// Background auto-fresh oracle (`[oracle]`). Opt-in; default OFF. When `auto_run` is enabled, the
-/// long-lived `rag-rat mcp` server runs the SCIP oracle for the active checkout when the index is
-/// stale and quiet, heavily throttled by two gates (a long quiet-period debounce + a
-/// minimum-interval floor) — see [`crate::index::oracle::auto_run_decision`]. SCIP production takes
-/// minutes while edits arrive in seconds, so edge-collapsing alone would thrash; both gates are
-/// required. Fail-open and detached: it never blocks a request and dies with the server process.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct OracleConfig {
-    /// Run the oracle in the background on the MCP server (default false — opt in explicitly).
-    pub auto_run: bool,
-    /// Run only after the index has been quiet (no change) for at least this long. The debounce
-    /// that keeps an active editing session from triggering a minutes-long SCIP pass on every
-    /// save.
-    pub auto_run_quiet_period_secs: u64,
-    /// And at most once this often, regardless of churn. The minimum-interval floor.
-    pub auto_run_min_interval_secs: u64,
-}
-
-impl Default for OracleConfig {
-    fn default() -> Self {
-        Self {
-            auto_run: false,
-            auto_run_quiet_period_secs: 900,
-            auto_run_min_interval_secs: 21_600,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct VersionCheckConfig {
-    /// Check crates.io for a newer published `rag-rat` and surface it to agents/operators (default
-    /// true). Opt out with `[version_check] enabled = false` in `rag-rat.toml`. The check is
-    /// best-effort, cached, and never blocks; disabling it makes no network calls at all.
-    pub enabled: bool,
-}
-
-impl Default for VersionCheckConfig {
-    fn default() -> Self {
-        Self { enabled: true }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct WatchConfig {
-    /// Run the background file watcher (default true). `RAG_RAT_NO_WATCH` overrides this off at
-    /// the call site.
-    pub enabled: bool,
-    /// Quiet window (ms) before a debounced reindex pass.
-    pub debounce_ms: u64,
-    /// Hard cap (ms): force a pass after this much continuous activity, so sustained writes never
-    /// starve the quiet-window debounce.
-    pub max_latency_ms: u64,
-    /// Periodic backstop: run a pass at least this often even with no events (0 disables). Covers
-    /// event-blind filesystems (NFS, WSL2 `/mnt`) and a watcher that missed events, and bounds how
-    /// long a wedged peer can leave the index stale.
-    pub periodic_sweep_secs: u64,
-}
-
-impl Default for WatchConfig {
-    fn default() -> Self {
-        Self { enabled: true, debounce_ms: 400, max_latency_ms: 2500, periodic_sweep_secs: 300 }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LogLevel {
-    Off,
-    Error,
-    Warn,
-    Info,
-    Debug,
-    Trace,
-}
-
-impl LogLevel {
-    fn parse(s: &str) -> Option<Self> {
-        Some(match s.trim().to_ascii_lowercase().as_str() {
-            "off" => Self::Off,
-            "error" => Self::Error,
-            "warn" => Self::Warn,
-            "info" => Self::Info,
-            "debug" => Self::Debug,
-            "trace" => Self::Trace,
-            _ => return None,
-        })
-    }
-
-    /// The EnvFilter directive string for this level.
-    pub fn as_filter_str(&self) -> &'static str {
-        match self {
-            Self::Off => "off",
-            Self::Error => "error",
-            Self::Warn => "warn",
-            Self::Info => "info",
-            Self::Debug => "debug",
-            Self::Trace => "trace",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LogFormat {
-    Text,
-    Json,
-}
-
-impl LogFormat {
-    fn parse(s: &str) -> Option<Self> {
-        match s.trim().to_ascii_lowercase().as_str() {
-            "text" => Some(Self::Text),
-            "json" => Some(Self::Json),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct LogConfig {
-    /// Master switch (default false). `RAG_RAT_LOG` can force-enable at init even when false.
-    pub enabled: bool,
-    pub level: LogLevel,
-    /// Optional per-subsystem EnvFilter directives (e.g. `rag_rat_core::index::ai=debug`).
-    pub filter: Option<String>,
-    /// Resolved absolute log dir (finalized in `load()` to the db sibling by default).
-    pub dir: PathBuf,
-    pub format: LogFormat,
-    /// Roll/prune a file once it exceeds this many bytes (0 disables the size check).
-    pub max_file_bytes: u64,
-    pub retention_days: u64,
-    pub max_files: u64,
-}
-
-impl Default for LogConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            level: LogLevel::Info,
-            filter: None,
-            dir: PathBuf::from(".rag-rat/logs"),
-            format: LogFormat::Text,
-            max_file_bytes: 50 * 1024 * 1024,
-            retention_days: 7,
-            max_files: 200,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct LlmConfig {
-    pub embedding: EmbeddingConfig,
-    pub dream: DreamLlmConfig,
-}
-
-/// Dream-mode model pass (`[llm.dream]`) — rag-rat's first generative-model dependency (#122),
-/// serving its verdict/compaction turns over an OpenAI-compatible chat endpoint. Mirrors
-/// [`EmbeddingConfig`]: an `enabled` gate plus a [`RemoteDreamConfig`] serving block (connect XOR
-/// ephemeral). Default OFF, so `rag-rat dream` stays 100% deterministic unless the operator opts
-/// in.
-///
-/// Unlike embeddings, `remote` is NOT optional: dream has no in-process backend, so an absent
-/// `[llm.dream.remote]` block still resolves to a serving config — [`RemoteDreamConfig::default`],
-/// a connect to a local Ollama (today's `[llm.dream.remote]` default).
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct DreamLlmConfig {
-    /// Run the model pass at all (default false — opt in explicitly). When false, `rag-rat dream
-    /// --verify` still runs the deterministic pass-0 findings; only the model turn is skipped.
-    pub enabled: bool,
-    /// Which chat server serves the dream turns (connect XOR ephemeral). Absent
-    /// `[llm.dream.remote]` → [`RemoteDreamConfig::default`] (a local-Ollama connect).
-    pub remote: RemoteDreamConfig,
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct EmbeddingConfig {
-    /// Which embedding backend to use for semantic (vector) recall. `init` picks a default based
-    /// on repo size; see [`EmbeddingBackend`].
-    pub backend: EmbeddingBackend,
-    pub runtime: EmbeddingRuntimeConfig,
-    /// Optional remote-embedding offload (`[llm.embedding.remote]`). When present, the
-    /// indexer can hand embedding work to an HTTP server (Ollama's `/api/embed`) instead of
-    /// running the model in-process — see [`RemoteEmbeddingConfig`]. Absent → `None` → in-process
-    /// embedding only. Parsed + validated here; the dispatch that consumes it lands in #317 task
-    /// 5.
-    pub remote: Option<RemoteEmbeddingConfig>,
-}
-
-/// The embedding backend selector (`[llm.embedding] model = "..."`).
-///
-/// Resolves the toml `model = "..."` string through the [`crate::embedding_models`] registry by its
-/// `model_id` (the HF path — NO aliases, #317), so ANY registered model is selectable by its full
-/// name — adding a model to the registry makes it selectable here with no edit. The embeddings-off
-/// choice (`none` / `off`) carries `None`. Kept a thin wrapper over a registry spec reference so
-/// `config` resolves model identity without depending on `index` (`crate::embedding_models` has no
-/// `index` dependency, so there is no cycle).
-///
-/// The common tiers, for reference:
-/// - `sentence-transformers/all-MiniLM-L6-v2` (the `EmbeddingBackend::default`): MiniLM transformer
-///   — best general quality, but the cold backfill is CPU-bound (~10-100 chunks/sec), so
-///   impractical for very large repos.
-/// - `minishlab/potion-retrieval-32M`: static token-vector lookup + mean-pool — ~100-500× faster on
-///   CPU at some retrieval-quality cost (no context/word-order). The choice for huge repos that
-///   still want vectors.
-/// - `none`: structural + BM25 only; no dense vectors. `semantic_search` degrades to BM25. The
-///   cheapest option for enormous codebases where any embedding backfill is too slow.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct EmbeddingBackend(Option<&'static EmbeddingModelSpec>);
-
-impl EmbeddingBackend {
-    /// Embeddings off — structural + BM25 only.
-    pub const NONE: Self = Self(None);
-
-    /// The FastEmbed MiniLM tier — the default backend (`init` recommends it for smaller repos).
-    /// Resolved from the registry by model_id.
-    pub fn fast_embed() -> Self {
-        Self(spec(FASTEMBED_MODEL_ID))
-    }
-
-    /// The Model2Vec static tier (`init` recommends it for very large repos).
-    pub fn model2vec() -> Self {
-        Self(spec(MODEL2VEC_MODEL_ID))
-    }
-
-    /// The toml selector for this backend — the model_id (the HF path `init` renders back into
-    /// `rag-rat.toml`), or `"none"` for the embeddings-off choice.
-    pub fn as_str(self) -> &'static str {
-        match self.0 {
-            Some(spec) => spec.model_id,
-            None => "none",
-        }
-    }
-
-    /// The persisted embedding-model id this backend installs/activates, or `None` for the
-    /// embeddings-off choice.
-    pub fn model_id(self) -> Option<&'static str> {
-        self.0.map(|spec| spec.model_id)
-    }
-
-    /// The registry [`Backend`] this selector resolves to (`None` for the embeddings-off choice).
-    /// Lets `config` reason about a `model = "..."` selection without re-deriving the mapping the
-    /// registry owns — e.g. the `[embedding.remote]` guardrail that rejects serving a
-    /// non-transformer (static/hash) model over Ollama.
-    pub fn registry_backend(self) -> Option<Backend> {
-        self.0.map(|spec| spec.backend)
-    }
-}
-
-impl Default for EmbeddingBackend {
-    fn default() -> Self {
-        Self::fast_embed()
-    }
-}
-
-impl FromStr for EmbeddingBackend {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        let value = value.trim();
-        // The embeddings-off keywords are case-insensitive; the MODEL selector is the HF path,
-        // which is CASE-SENSITIVE (`all-MiniLM-L6-v2`, `BAAI/...`) — match it verbatim via
-        // `spec`, never lowercased, or the case-sensitive model_id lookup would miss.
-        match value.to_ascii_lowercase().as_str() {
-            "none" | "off" | "bm25" => Ok(Self::NONE),
-            _ => spec(value)
-                .map(|spec| Self(Some(spec)))
-                .ok_or_else(|| ConfigError::UnknownEmbeddingBackend(value.to_string())),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct EmbeddingRuntimeConfig {
-    pub batch_size: u32,
-    pub ort_threads: Option<u32>,
-    pub omp_threads: Option<u32>,
-    pub max_embedding_chars: usize,
-}
-
-impl Default for EmbeddingRuntimeConfig {
-    fn default() -> Self {
-        Self {
-            batch_size: 64,
-            ort_threads: Some(4),
-            omp_threads: Some(1),
-            max_embedding_chars: 4000,
-        }
-    }
-}
-
-/// Which OpenAI-compatible embedding server serves a `[remote]` block. All three speak the SAME
-/// wire API (`POST /v1/embeddings`), so the embedding client is IDENTICAL regardless of backend —
-/// the selector only routes ephemeral provisioning (which cookbook container), the freshness /
-/// vector- identity marker (different backends can produce slightly different vectors), the tune
-/// cache key, and the `ai_models.runtime` install marker.
-#[derive(
-    Debug,
-    Clone,
-    Copy,
-    PartialEq,
-    Eq,
-    Default,
-    Serialize,
-    Deserialize,
-    strum::EnumString,
-    strum::IntoStaticStr,
-)]
-#[serde(rename_all = "lowercase")]
-#[strum(serialize_all = "lowercase", ascii_case_insensitive)]
-pub enum RemoteBackend {
-    /// ollama, via its `/v1/embeddings` OpenAI-compatibility route (the default; back-compat).
-    #[default]
-    Ollama,
-    /// michaelfeil/infinity (`infinity_emb v2 --model-id <hf>`).
-    Infinity,
-    /// vLLM in embedding mode (`vllm serve <hf> --runner pooling --host 0.0.0.0`).
-    Vllm,
-}
-
-impl RemoteBackend {
-    /// The stable wire/DB string (matches the serde repr): the `ai_models.runtime` marker, and the
-    /// backend discriminator folded into the freshness key + the tune cache key.
-    pub fn as_db_str(self) -> &'static str {
-        self.into()
-    }
-
-    /// Parse a config `backend = "..."` value (case-insensitive). `None` for an unknown value — the
-    /// config layer turns that into `ConfigError::RemoteBackendUnknown`.
-    pub fn from_db_str(s: &str) -> Option<Self> {
-        s.trim().parse().ok()
-    }
-
-    /// The HTTP path (appended to the endpoint) of this backend's embeddings route. The
-    /// request/response SHAPE is identical across backends (OpenAI `{model,input}` →
-    /// `{data:[{embedding,index}]}`); only the PATH differs: ollama and vLLM expose the
-    /// OpenAI-standard `/v1/embeddings`, while michaelfeil/infinity's `v2` server serves it at
-    /// `/embeddings` (verified live against `michaelf34/infinity:latest-cpu`).
-    pub fn embed_path(self) -> &'static str {
-        match self {
-            RemoteBackend::Ollama | RemoteBackend::Vllm => "/v1/embeddings",
-            RemoteBackend::Infinity => "/embeddings",
-        }
-    }
-
-    /// Whether this backend can serve CHAT completions (the dream verdict/compaction turn). Ollama
-    /// (same image, pull a chat model) and vLLM (a generation runner, no `--runner pooling`) both
-    /// can; michaelfeil/infinity is embeddings/rerank/classify only, so a `[llm.dream.remote]`
-    /// block on `infinity` is a config error (`DreamBackendCannotServeChat`). Embeddings request
-    /// the `embed` capability; dream requests `chat`.
-    pub fn supports_chat(self) -> bool {
-        match self {
-            RemoteBackend::Ollama | RemoteBackend::Vllm => true,
-            RemoteBackend::Infinity => false,
-        }
-    }
-
-    /// The HTTP path (appended to the endpoint) of this backend's chat-completions route. Uniform
-    /// across chat-capable backends — ollama and vLLM both expose the OpenAI-standard
-    /// `/v1/chat/completions`; only the SERVING differs (see `supports_chat`), not the route.
-    pub fn chat_path(self) -> &'static str {
-        "/v1/chat/completions"
-    }
-
-    /// How long the ephemeral cookbook may take to provision + serve a box for this backend before
-    /// rag-rat gives up — BOTH the Rust-side handshake deadline and (minus a margin) the
-    /// `provision_timeout_s` the recipe budgets against.
-    ///
-    /// vLLM's published image (`vllm/vllm-openai`, a ~10–15 GB CUDA image) is an order of magnitude
-    /// larger than ollama's or infinity's, so its cold start (image pull + GPU model load) needs a
-    /// much longer ceiling: 300s times it out on Modal (the create alone can eat most of it,
-    /// leaving too little for `waitUntilReady`). ollama/infinity cold-start in well under 300s.
-    pub fn provision_timeout(self) -> Duration {
-        match self {
-            RemoteBackend::Ollama | RemoteBackend::Infinity => Duration::from_secs(300),
-            RemoteBackend::Vllm => Duration::from_secs(900),
-        }
-    }
-}
-
-/// Remote-embedding offload (`[llm.embedding.remote]`). Hands embedding work to an
-/// OpenAI-compatible HTTP server (`POST /v1/embeddings` on ollama/infinity/vLLM) instead of running
-/// the model in-process — the lever for huge repos whose in-process backfill is too slow on the
-/// indexing box. Optional: absent → in-process embedding only.
-///
-/// Carries NO `dim` field: the vector dimension comes from the registry spec of the SELECTED model
-/// (`model = "sentence-transformers/all-MiniLM-L6-v2"`, dim 384) and is validated against the
-/// server's first response at runtime by the embedder; duplicating it here would be redundant and
-/// drift-prone. The `backend` selector (default `ollama`) picks WHICH server; the runtime is
-/// implied by this block's mere PRESENCE (#317 rework).
-///
-/// The mode is INFERRED, not configured — there is no `mode` field (#318). EXACTLY ONE of
-/// `endpoint` / `cookbook` is set:
-/// - `endpoint` present → CONNECT: talk to an already-running Ollama at that URL.
-/// - `cookbook` present → EPHEMERAL: the bulk-`reconcile` path provisions an on-demand box via the
-///   cookbook subprocess, embeds the repo against it, then tears it down. Queries use the LOCAL box
-///   at `query_endpoint` (same model → same GGUF vector space as the remote-embedded chunks).
-///
-/// `Serialize` lets the install step persist this verbatim into the secret-free
-/// `active_embedding_remote_config` meta, so the `conn`-based `active_embedder` can reconstruct the
-/// remote embedder for both chunk-embed (reconcile) and query-embed (search) without threading
-/// config through the search path. SECRET-FREE: `auth_env` is the env-var NAME, never a token, so
-/// the serialized JSON holds no secret. `Deserialize` is the read side of that meta round-trip.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RemoteEmbeddingConfig {
-    /// The SERVER-side model name sent in the `/v1/embeddings` request body — the server's own
-    /// identifier, NOT a `rag-rat` registry alias (the registry only supplies the dim parity
-    /// contract). For `backend = "ollama"` this is the ollama model name (e.g. `"all-minilm"`);
-    /// for infinity/vLLM it is the HuggingFace id the server was launched with (e.g.
-    /// `"sentence-transformers/all-MiniLM-L6-v2"`).
-    pub model: String,
-    /// Which OpenAI-compatible server serves this block (`ollama` | `infinity` | `vllm`). The wire
-    /// call is identical across backends; this only routes provisioning + the
-    /// freshness/tune/install markers. `#[serde(default)]` → `Ollama` so config JSON persisted
-    /// before this field (and any omitted TOML) still deserializes as the pre-existing ollama
-    /// behavior.
-    #[serde(default)]
-    pub backend: RemoteBackend,
-    /// CONNECT: base URL of an already-running Ollama (e.g. `"http://localhost:11434"`);
-    /// `/api/embed` is appended by the embedder. Mutually exclusive with `cookbook`.
-    pub endpoint: Option<String>,
-    /// EPHEMERAL: the cookbook recipe rag-rat spawns to provision an on-demand box — an npm
-    /// package spec (`"@rag-rat/cookbook/modal"`, run via `npx -y`) or a recipe file path
-    /// (`.mjs`/`.js` → `node`, `.ts` → `npx tsx`). Mutually exclusive with `endpoint`.
-    pub cookbook: Option<String>,
-    /// EPHEMERAL: the LOCAL server used for QUERY embedding after the box is torn down (queries
-    /// embed the same model on the same `backend` → identical vector space). Defaults to
-    /// `http://localhost:11434` (local Ollama) ONLY for `backend = ollama`; a non-ollama backend
-    /// must set this explicitly (its route/model differ, so the Ollama default would silently
-    /// break query embedding — see `ConfigError::RemoteQueryEndpointRequiredForBackend`).
-    /// Ignored in connect mode.
-    pub query_endpoint: Option<String>,
-    /// Name of the environment variable holding the bearer token, if the server needs auth. Local
-    /// Ollama needs none, so this is optional; the embedder reads the var once at construction.
-    pub auth_env: Option<String>,
-    /// EPHEMERAL-only: the GPU the cookbook recipe should provision for the on-demand box. The
-    /// value is PROVIDER-specific and validated by the provider at provision time, not here —
-    /// Modal wants a GPU class (`A10G`/`T4`/`A100`), RunPod a `gpuTypeId`. `None` lets the
-    /// recipe pick its default (Modal defaults to CPU; RunPod to a cheap on-demand GPU). Set
-    /// together with a connect `endpoint` it is a config error (`RemoteGpuRequiresCookbook`).
-    ///
-    /// `#[serde(default)]`: this field post-dates the meta round-trip, so a config JSON persisted
-    /// by an older binary (no `gpu` key) must still deserialize (→ `None`) instead of
-    /// erroring.
-    #[serde(default)]
-    pub gpu: Option<String>,
-    /// Optional Ollama context window (`options.num_ctx`) for `/api/embed` requests. Some local
-    /// embedding GGUFs default too small for rag-rat's code chunks; this lets config raise the
-    /// server context without shrinking chunk size or batch throughput.
-    ///
-    /// `#[serde(default)]`: this field post-dates the meta round-trip, so a config JSON persisted
-    /// by an older binary (no `num_ctx` key) must still deserialize (→ `None`) instead of
-    /// erroring.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub num_ctx: Option<u32>,
-    /// How many texts to send per `/api/embed` request.
-    pub batch_size: u32,
-    /// How many `/api/embed` requests the remote embedder may keep in flight.
-    ///
-    /// `#[serde(default)]`: this field post-dates the meta round-trip. Older persisted remote meta
-    /// has no mode-aware TOML context, so deserialize it as single-flight instead of making an
-    /// existing connect endpoint fan out 32 requests after upgrade.
-    #[serde(default = "legacy_remote_embedding_concurrency")]
-    pub concurrency: u32,
-    /// Maximum total input characters to include in one `/api/embed` request.
-    ///
-    /// `#[serde(default)]`: this field post-dates the meta round-trip, so a config JSON persisted
-    /// by an older binary (no `max_batch_chars` key) must still deserialize with the default
-    /// instead of erroring.
-    #[serde(default = "default_remote_embedding_max_batch_chars")]
-    pub max_batch_chars: usize,
-    /// Per-request HTTP timeout, in seconds.
-    pub request_timeout_s: u64,
-}
-
-/// The default LOCAL Ollama URL for ephemeral query embedding when `query_endpoint` is omitted.
-pub const DEFAULT_QUERY_ENDPOINT: &str = "http://localhost:11434";
-/// Upper bound a user may set for `[remote] concurrency`. Headroom for GPU backends
-/// (infinity/vLLM), which do real dynamic-batched inference and keep scaling with client fan-out
-/// well past 128 (an L4 serving all-MiniLM via infinity was still climbing at 128 with zero
-/// failures). ollama is server-bound (a higher value doesn't help it), so this is opt-in per
-/// config; the defaults stay low. The auto-tuner sweeps within whatever cap the user sets and
-/// clamps the knee to it.
-pub const MAX_REMOTE_EMBEDDING_CONCURRENCY: u32 = 512;
-const DEFAULT_REMOTE_EMBEDDING_CONCURRENCY: u32 = 32;
-const LEGACY_REMOTE_EMBEDDING_CONCURRENCY: u32 = 1;
-const DEFAULT_REMOTE_EMBEDDING_MAX_BATCH_CHARS: usize = 384_000;
-
-fn legacy_remote_embedding_concurrency() -> u32 {
-    LEGACY_REMOTE_EMBEDDING_CONCURRENCY
-}
-
-fn default_remote_embedding_max_batch_chars() -> usize {
-    DEFAULT_REMOTE_EMBEDDING_MAX_BATCH_CHARS
-}
-
-impl Default for RemoteEmbeddingConfig {
-    fn default() -> Self {
-        Self {
-            model: String::new(),
-            backend: RemoteBackend::Ollama,
-            endpoint: None,
-            cookbook: None,
-            query_endpoint: None,
-            auth_env: None,
-            gpu: None,
-            num_ctx: None,
-            batch_size: 256,
-            concurrency: DEFAULT_REMOTE_EMBEDDING_CONCURRENCY,
-            max_batch_chars: DEFAULT_REMOTE_EMBEDDING_MAX_BATCH_CHARS,
-            request_timeout_s: 60,
-        }
-    }
-}
-
-impl RemoteEmbeddingConfig {
-    /// Default for a missing TOML `concurrency` field after the mode has been inferred.
-    ///
-    /// Connect-mode configs may point at ordinary Ollama servers that were not started with
-    /// matching `OLLAMA_NUM_PARALLEL`, so omitted connect concurrency stays single-flight. Cookbook
-    /// configs are provisioned by rag-rat recipes that align server-side parallelism, so omitted
-    /// ephemeral configs use the new parallel default.
-    pub fn omitted_concurrency_default(is_connect: bool) -> u32 {
-        if is_connect {
-            LEGACY_REMOTE_EMBEDDING_CONCURRENCY
-        } else {
-            DEFAULT_REMOTE_EMBEDDING_CONCURRENCY
-        }
-    }
-
-    /// Runtime-safe concurrency for deserialized/persisted configs. TOML parsing rejects explicit
-    /// values above [`MAX_REMOTE_EMBEDDING_CONCURRENCY`], but older or manually-edited DB metadata
-    /// can bypass that validation.
-    pub fn bounded_concurrency(&self) -> u32 {
-        Self::bounded_concurrency_value(self.concurrency)
-    }
-
-    pub fn bounded_concurrency_value(value: u32) -> u32 {
-        value.clamp(1, MAX_REMOTE_EMBEDDING_CONCURRENCY)
-    }
-
-    /// CONNECT mode: an already-running server at `endpoint`. Exactly one of connect/ephemeral
-    /// holds (config validation guarantees it), so `is_connect() == !is_ephemeral()`.
-    pub fn is_connect(&self) -> bool {
-        self.endpoint.is_some()
-    }
-
-    /// EPHEMERAL mode: provision an on-demand box via `cookbook` for the bulk reconcile.
-    pub fn is_ephemeral(&self) -> bool {
-        self.cookbook.is_some()
-    }
-}
-
-/// Remote-serving config for the dream model pass (`[llm.dream.remote]`). The CHAT-flavored mirror
-/// of [`RemoteEmbeddingConfig`]: it hands the verdict/compaction turn to an OpenAI-compatible chat
-/// server (`POST /v1/chat/completions` on ollama/vLLM) instead of running a model in-process.
-///
-/// Chat is one-turn and budget-capped, so this carries NONE of the embedding block's batching knobs
-/// (`query_endpoint`, `num_ctx`, `batch_size`, `concurrency`, `max_batch_chars`) — only the serving
-/// selector plus a request timeout.
-///
-/// Like embeddings, the mode is INFERRED, not configured — EXACTLY ONE of `endpoint` / `cookbook`
-/// is set:
-/// - `endpoint` present → CONNECT: talk to an already-running chat server at that URL.
-/// - `cookbook` present → EPHEMERAL: the dream command provisions an on-demand box via the cookbook
-///   subprocess, runs the pass against it, then tears it down.
-///
-/// NOT `Serialize`/`Deserialize`: dream config does not round-trip through the index meta (unlike
-/// the embedder, which reconstructs itself from persisted config), matching the old
-/// `DreamModelConfig`.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct RemoteDreamConfig {
-    /// Which OpenAI-compatible server serves this block (`ollama` | `vllm`). `infinity` is
-    /// embed-only and rejected at parse time (`DreamBackendCannotServeChat`). Omitted → `Ollama`.
-    pub backend: RemoteBackend,
-    /// CONNECT: base URL of an already-running chat server (e.g. `"http://localhost:11434"`);
-    /// `/v1/chat/completions` is appended by the verdict client. Mutually exclusive with
-    /// `cookbook`.
-    pub endpoint: Option<String>,
-    /// EPHEMERAL: the cookbook recipe rag-rat spawns to provision an on-demand box — an npm
-    /// package spec (`"@rag-rat/cookbook modal"`, run via `npx -y`) or a recipe file path.
-    /// Mutually exclusive with `endpoint`.
-    pub cookbook: Option<String>,
-    /// The SERVER-side chat model sent in the request body — the server's own identifier, NOT a
-    /// rag-rat registry alias. For `backend = "ollama"` the ollama model name (e.g. `"qwen3:8b"`);
-    /// for vLLM the HuggingFace id the server was launched with (e.g. `"Qwen/Qwen3-4B-Instruct"`).
-    pub model: String,
-    /// EPHEMERAL-only: the GPU the cookbook recipe should provision. PROVIDER-specific and
-    /// validated by the provider at provision time, not here. Set together with a connect
-    /// `endpoint` it is a config error (`DreamRemoteGpuRequiresCookbook`).
-    pub gpu: Option<String>,
-    /// Name of the environment variable holding the bearer token, if the server needs auth. Local
-    /// Ollama needs none, so this is optional; the verdict client reads the var once at
-    /// construction.
-    pub auth_env: Option<String>,
-    /// Per-request HTTP timeout, in seconds. A dense evidence pack against a remote model can take
-    /// a while, so the default is generous.
-    pub request_timeout_s: u64,
-}
-
-impl Default for RemoteDreamConfig {
-    /// A local-Ollama CONNECT — byte-for-byte the pre-migration `[dream.model]` default, so an
-    /// absent `[llm.dream.remote]` block preserves today's behavior.
-    fn default() -> Self {
-        Self {
-            backend: RemoteBackend::Ollama,
-            endpoint: Some("http://localhost:11434".to_string()),
-            cookbook: None,
-            model: "qwen3:4b-instruct".to_string(),
-            gpu: None,
-            auth_env: None,
-            request_timeout_s: 300,
-        }
-    }
-}
-
-impl RemoteDreamConfig {
-    /// CONNECT mode: an already-running server at `endpoint`. Exactly one of connect/ephemeral
-    /// holds (config validation guarantees it), so `is_connect() == !is_ephemeral()`.
-    pub fn is_connect(&self) -> bool {
-        self.endpoint.is_some()
-    }
-
-    /// EPHEMERAL mode: provision an on-demand box via `cookbook` for the dream pass.
-    pub fn is_ephemeral(&self) -> bool {
-        self.cookbook.is_some()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ResolvedTarget {
-    pub name: String,
-    pub language: Language,
-    pub directories: Vec<PathBuf>,
-    pub include: Vec<String>,
-    pub exclude: Vec<String>,
-    pub kind: TargetKind,
-}
-
-impl ResolvedTarget {
-    /// Indexing precedence when more than one target claims the same file (lower sorts first /
-    /// wins). Two levels: (1) kind — generated/tests/docs before source, the existing order;
-    /// (2) within a kind, a language that claims an ambiguous extension as an upgrade
-    /// ([`Language::upgrades_ambiguous_extension`]) wins it — so a `.h` covered by both a `c` and a
-    /// `cpp` binding indexes as C++ (the deliberate intent), not C (alphabetical accident). Both
-    /// the full-rebuild walk (first-claimer-wins) and the incremental per-path resolver use
-    /// this.
-    pub fn index_precedence(&self) -> (u8, u8) {
-        let kind_rank = match self.kind {
-            TargetKind::Generated => 0,
-            TargetKind::Tests => 1,
-            TargetKind::Docs => 2,
-            TargetKind::Source => 3,
-        };
-        let upgrade_rank = u8::from(!self.language.upgrades_ambiguous_extension());
-        (kind_rank, upgrade_rank)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TargetKind {
-    Source,
-    Generated,
-    Docs,
-    Tests,
-}
-
-impl TargetKind {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Source => "source",
-            Self::Generated => "generated",
-            Self::Docs => "docs",
-            Self::Tests => "tests",
-        }
-    }
-}
-
-impl FromStr for TargetKind {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_ascii_lowercase().as_str() {
-            "source" => Ok(Self::Source),
-            "generated" => Ok(Self::Generated),
-            "docs" => Ok(Self::Docs),
-            "tests" | "test" => Ok(Self::Tests),
-            other => Err(ConfigError::UnknownTargetKind(other.to_string())),
-        }
-    }
-}
-
-impl Config {
-    /// The deduplicated set of target directories (relative to [`Config::root`]) across all
-    /// targets, in stable order. Used to scope `.gitignore` nested-discovery to the indexed trees
-    /// (see [`crate::index::ignore_rules::IgnoreMatcher::compile`]) instead of recursing the whole
-    /// root into unindexed siblings.
-    pub fn target_directories(&self) -> Vec<PathBuf> {
-        let mut seen = BTreeSet::new();
-        let mut dirs = Vec::new();
-        for target in &self.targets {
-            for dir in &target.directories {
-                if seen.insert(dir.clone()) {
-                    dirs.push(dir.clone());
-                }
-            }
-        }
-        dirs
-    }
-
-    /// A copy of this (BASE) config whose `targets` are re-resolved from a LINKED worktree's own
-    /// `rag-rat.toml` (at `<linked_worktree_root>/rag-rat.toml`), so an overlay refresh indexes
-    /// that branch with its OWN target set — not the sweeping process's. The shared `root`
-    /// (anchored to main), `database`, and the rest are kept: the overlay still resolves
-    /// against the one base index, but the delta + ignore filtering see the branch's targets.
-    /// Returns the config unchanged when the linked worktree has no readable/valid
-    /// `rag-rat.toml`, or when its targets don't validate against the linked checkout — so a
-    /// malformed branch config degrades to base targets rather than dropping the worktree. The
-    /// branch targets are root-relative, so they apply to the shared `root` directly (#219
-    /// review).
-    ///
-    /// Used by `refresh_worktree_overlays`: the main watcher/maintenance process refreshes every
-    /// linked worktree, but a worktree whose branch ADDS a target (e.g. `extra/`) would otherwise
-    /// be filtered against the sweeper's targets — pruning overlay rows a branch-launched hook
-    /// indexed.
-    pub fn for_linked_worktree_overlay(&self, linked_path: &Path) -> Self {
-        let linked_targets = (|| {
-            // `linked_path` may be the checkout root, a subdir of it, or the git dir (a hook); the
-            // branch `rag-rat.toml` lives at the WORKDIR top, so resolve the workdir first.
-            let workdir = crate::index::discover_repo(linked_path)
-                .ok()
-                .and_then(|repo| repo.workdir().map(Path::to_path_buf))
-                .unwrap_or_else(|| linked_path.to_path_buf());
-            let text = fs::read_to_string(workdir.join("rag-rat.toml")).ok()?;
-            let raw: RawConfig = toml::from_str(&text).ok()?;
-            // Targets are relative to the config's `[index].root` (a subdir layout puts the toml at
-            // the worktree top with `root = "<subdir>"`); resolve + validate them there so the
-            // stored, root-relative directories match the base config's spelling exactly.
-            let target_root = workdir.join(raw.index.root.as_deref().unwrap_or("."));
-            resolve_targets(&target_root, raw.target_bindings, raw.target).ok()
-        })();
-        match linked_targets {
-            Some(targets) => Self { targets, ..self.clone() },
-            None => self.clone(),
-        }
-    }
-
-    /// Load a config, resolving WHICH `rag-rat.toml` GOVERNS at one seam: in a linked git
-    /// worktree, the MAIN worktree's config file is authoritative for the WHOLE config — identity,
-    /// database location, targets, models, everything. A branch-local `rag-rat.toml` (an older
-    /// branch, a divergent checkout) cannot fork any of it; when its content differs from main's
-    /// it is ignored with a one-line warning naming the ignored file. This subsumes the historical
-    /// per-key anchoring (root #218/#219, targets #219, `repo_id` #413, `database` A7) — the
-    /// question "which config governs this repo" is answered once, so new keys are main-anchored
-    /// by default and cannot re-open the split-brain class.
-    ///
-    /// DECLARED WORKTREE-LOCAL ALLOW-LIST (the only keys that legitimately vary per worktree):
-    ///  * `target_bindings` / `[[target]]` — FOR THE OVERLAY INDEX ONLY, read by
-    ///    [`Config::for_linked_worktree_overlay`] from the linked checkout's own file, because a
-    ///    branch may add/remove source dirs and its overlay must index its own file set (#219). The
-    ///    BASE config's targets remain main-anchored here.
-    ///
-    /// Everything else, present and future, resolves from the governing (main) config.
-    ///
-    /// EDGE POSTURES:
-    ///  * Main worktree resolvable but CONFIG-LESS → the local config governs, with a warning (the
-    ///    repo's config belongs in main; until it exists there, the branch copy is all we have —
-    ///    best-effort, mirroring the old per-key fallbacks).
-    ///  * Main config exists but FAILS to parse/read → the error PROPAGATES: loading from any
-    ///    worktree must behave like loading from main, errors included (silently falling back to
-    ///    the branch config would fork the repo exactly when main is briefly broken).
-    ///  * No resolvable main (bare-repo hubs, pruned main, custom GIT_DIR, non-git roots) →
-    ///    `main_worktree_root` is `None`, the local config governs unchanged — there is no
-    ///    designated main to defer to.
-    pub fn load(path: impl AsRef<Path>) -> Result<Self, ConfigError> {
-        let path = path.as_ref();
-        let text = fs::read_to_string(path)?;
-        // Parse the LOCAL file but DO NOT fail yet: when a main config governs, the branch-local
-        // file's contents are irrelevant by design — a parse/validation failure there must fold
-        // into the divergence warning, never block every command from the linked checkout (Codex
-        // batch 8, finding 2). Wherever the local config GOVERNS, the error is fatal as always.
-        // The `[local_ai]` / `[dream]` rejections are part of local validity — see the
-        // `RawConfig` presence-capture fields.
-        let local_parse: Result<RawConfig, ConfigError> =
-            toml::from_str::<RawConfig>(&text).map_err(ConfigError::from).and_then(|raw| {
-                if raw.local_ai.is_some() {
-                    Err(ConfigError::LocalAiTableRenamed)
-                } else if raw.dream.is_some() {
-                    Err(ConfigError::DreamTableMoved)
-                } else {
-                    Ok(raw)
-                }
-            });
-        let local_config_dir = path.parent().unwrap_or_else(|| Path::new("."));
-        // The topology subject must be a discoverable directory: a RELATIVE config path like
-        // `rag-rat.toml` has the EMPTY path as its parent (`Path::parent` yields `Some("")`, not
-        // `None`), which git discovery cannot open — it means the process cwd.
-        let local_checkout =
-            if local_config_dir.as_os_str().is_empty() { Path::new(".") } else { local_config_dir };
-
-        // Best-effort resolution of what the LOCAL checkout's own `[index] root` names, taken
-        // BEFORE the governing seam below picks a winner — used only to detect + report a re-anchor
-        // to the caller (#427), never to decide anything (the seam is the sole source of truth for
-        // governance). `None` on any parse/resolution failure; that is not a second error path,
-        // just a diagnostic that stays silent when it cannot be computed.
-        let local_root_named: Option<PathBuf> = local_parse.as_ref().ok().and_then(|local_raw| {
-            normalize_existing_dir(
-                &local_config_dir
-                    .join(local_raw.index.root.clone().unwrap_or_else(|| ".".to_string())),
-            )
-            .ok()
-        });
-
-        // THE GOVERNING SEAM (see the doc comment). Linked-ness comes from git TOPOLOGY — the
-        // checkout holding the config file vs the repo's designated main worktree
-        // ([`linked_worktree_main_root`]) — and governance is UNCONDITIONAL on that predicate.
-        // It must never hang off a root-anchoring proxy: a branch-only `[index] root` makes
-        // `anchor_root_to_main_worktree` return the local root unchanged, and an equality trigger
-        // would then let the branch config govern database/identity/models — the exact
-        // split-brain the seam exists to prevent (Codex batch 8, finding 3). Anchoring outcomes
-        // affect ROOT resolution only, never who governs.
-        let (mut raw, config_dir, root, target_validation_root) =
-            match linked_worktree_main_root(local_checkout) {
-                Some(main_top) => match governing_main_config(&main_top)? {
-                    Some((main_raw, main_config_dir)) => {
-                        let divergent = match &local_parse {
-                            Ok(local_raw) => *local_raw != main_raw,
-                            Err(_) => true,
-                        };
-                        if divergent {
-                            let ignored =
-                                path.canonicalize().unwrap_or_else(|_| path.to_path_buf());
-                            let invalid_note =
-                                if local_parse.is_err() { " (also invalid)" } else { "" };
-                            eprintln!(
-                                "rag-rat: ignoring branch config{invalid_note} {} — in a linked \
-                                 worktree the main worktree's config governs ({}); edit that file \
-                                 instead",
-                                ignored.display(),
-                                main_config_dir.join("rag-rat.toml").display(),
-                            );
-                        }
-                        // Re-derive root from MAIN's own config, exactly as loading it directly
-                        // would (its root is already the main worktree — anchoring is identity).
-                        let main_root =
-                            normalize_existing_dir(&main_config_dir.join(
-                                main_raw.index.root.clone().unwrap_or_else(|| ".".to_string()),
-                            ))?;
-                        (main_raw, main_config_dir, main_root.clone(), main_root)
-                    },
-                    None => {
-                        // Config-less main: the LOCAL config governs (best-effort, loudly), so
-                        // its validity is fatal exactly as in a non-linked checkout.
-                        let local_raw = local_parse?;
-                        eprintln!(
-                            "rag-rat: the main worktree has no rag-rat.toml; using {} until one \
-                             exists there (the repo's config belongs in the main worktree)",
-                            path.display(),
-                        );
-                        // Root stays anchored so the shared index still keys off the main
-                        // checkout; targets validate against the local checkout where they
-                        // exist (#219).
-                        let local_root = normalize_existing_dir(&local_config_dir.join(
-                            local_raw.index.root.clone().unwrap_or_else(|| ".".to_string()),
-                        ))?;
-                        let anchored_root = anchor_root_to_main_worktree(&local_root);
-                        (local_raw, local_config_dir.to_path_buf(), anchored_root, local_root)
-                    },
-                },
-                None => {
-                    // The config's own checkout is the main worktree (or there is no designated
-                    // main): the local config governs and its validity is fatal. Root anchoring
-                    // still applies for the exotic `[index] root` pointing into a linked
-                    // checkout (#218/#219) — an anchoring concern, not a governance one.
-                    let local_raw = local_parse?;
-                    let local_root = normalize_existing_dir(
-                        &local_config_dir
-                            .join(local_raw.index.root.clone().unwrap_or_else(|| ".".to_string())),
-                    )?;
-                    let anchored_root = anchor_root_to_main_worktree(&local_root);
-                    (local_raw, local_config_dir.to_path_buf(), anchored_root, local_root)
-                },
-            };
-
-        // #427: `root` may have ended up different from what the LOCAL checkout's own `[index]
-        // root` names — either because a linked worktree's config was overridden wholesale by
-        // MAIN's (the common case, above), or because `anchor_root_to_main_worktree` rebased an
-        // exotic branch-only root (#218/#219). Either way, report the pre-anchor value so `index`
-        // can warn the operator they're indexing the main checkout, not the worktree they named.
-        let source_root_reanchored_from = local_root_named.filter(|named| *named != root);
-
-        // The database path (A7 default flip): an explicit `database` key is honored as-is — the
-        // deprecated per-repo deployment, that repo stays un-consolidated and never syncs. ABSENT,
-        // the default is the CONSOLIDATED GLOBAL store, EXCEPT a pre-existing legacy
-        // `.rag-rat/index.sqlite` is kept (with a deprecation nudge toward `rag-rat consolidate`).
-        // Relative explicit paths (and the legacy path) resolve against the MAIN worktree TOP —
-        // NOT `root`, which may be a subdirectory — so every worktree of a repo AND any
-        // `root="<subdir>"` config land on the SAME index.
-        let db_base = main_worktree_root(&root).unwrap_or_else(|| root.clone());
-        let repo_id_override =
-            raw.index.repo_id.take().map(|id| id.trim().to_string()).filter(|id| !id.is_empty());
-        let governing_database_key = raw.index.database.take();
-        let database_key_pinned = governing_database_key.is_some();
-        let database = match governing_database_key {
-            Some(db) if Path::new(&db).is_absolute() => PathBuf::from(db),
-            Some(db) => db_base.join(db),
-            // The keyless default probes the repo IDENTITY (root + the governing `[index]
-            // repo_id` pin): only an identity-BEARING root may land in the shared global store —
-            // see `default_database_with_disposition`.
-            None => resolve_default_database(&db_base, &root, repo_id_override.as_deref()),
-        };
-        // The identity gate's SECOND entrance (Codex batch 8, finding 5): an explicit pin AT the
-        // consolidated global store bypasses the keyless identity gate above, and an
-        // identity-less root (non-git, unborn HEAD) opening the shared store would fall through
-        // to adoption's sole-repo fallback — scoping this project onto whichever SIBLING repo
-        // sorts first. Refuse at resolution with the remedy. (The structural backstop for every
-        // other shared-path pin shape lives in `adopt_repo_from_config`: an identity-less open
-        // never sole-picks on a multi-repo database.)
-        if database_key_pinned
-            && Some(database.as_path()) == crate::data_dir::global_database_path().as_deref()
-            && !crate::repo_identity::identity_is_resolvable(&root, repo_id_override.as_deref())
-        {
-            return Err(ConfigError::GlobalPinWithoutIdentity);
-        }
-        // Targets resolve from the GOVERNING config; validation runs against the checkout that
-        // config describes (main when main governs; the local checkout on the config-less-main
-        // fallback, tolerating branch-only dirs — #219). `ResolvedTarget.directories` are
-        // root-relative, so the stored targets are checkout-independent either way. A linked
-        // branch's own target set is NOT lost: the overlay refresh reads the branch config via
-        // `for_linked_worktree_overlay` and indexes the branch with it (#219).
-        let targets = resolve_targets(&target_validation_root, raw.target_bindings, raw.target)?;
-        let mut llm = LlmConfig::try_from(raw.llm)?;
-        // Resolve a RELATIVE cookbook recipe PATH against the GOVERNING config dir, not the
-        // process CWD (R6): the recipe is handed to `node`/`npx`, which resolve it against
-        // wherever reconcile/the watcher runs — ENOENT from a subdir or a daemon.
-        if let Some(remote) = llm.embedding.remote.as_mut()
-            && let Some(cookbook) = remote.cookbook.as_ref()
-            && let Some(resolved) = resolve_relative_cookbook_path(cookbook, &config_dir)
-        {
-            remote.cookbook = Some(resolved);
-        }
-        // Same relative-cookbook resolution for the dream remote — its recipe is handed to
-        // `node`/`npx` too, so a relative path must resolve against the config dir, not the process
-        // CWD. `remote` is not optional for dream (a local-Ollama connect default), so only the
-        // ephemeral case has a cookbook to rewrite.
-        if let Some(cookbook) = llm.dream.remote.cookbook.as_ref()
-            && let Some(resolved) = resolve_relative_cookbook_path(cookbook, &config_dir)
-        {
-            llm.dream.remote.cookbook = Some(resolved);
-        }
-        let watch = raw.watch.into();
-        let version_check = raw.version_check.into();
-        let oracle = raw.oracle.into();
-        let search = raw.search.into();
-        let memory = MemoryConfig::try_from(raw.memory)?;
-        let mut log = LogConfig::try_from(raw.log)?;
-        // Finalize `dir`: empty (unset) → sibling of the db (`<db_parent>/logs`); a set value is
-        // resolved relative to the GOVERNING config dir (absolute honored).
-        log.dir = if log.dir.as_os_str().is_empty() {
-            database.parent().map(|p| p.join("logs")).unwrap_or_else(|| PathBuf::from("logs"))
-        } else if log.dir.is_absolute() {
-            log.dir.clone()
-        } else {
-            config_dir.join(&log.dir)
-        };
-
-        Ok(Self {
-            root,
-            database,
-            targets,
-            llm,
-            watch,
-            version_check,
-            oracle,
-            search,
-            memory,
-            log,
-            repo_id_override,
-            database_key_pinned,
-            source_root_reanchored_from,
-            allow_empty: false,
-        })
-    }
-}
-
-/// The MAIN worktree's parsed config, when one exists — the governing side of `Config::load`'s
-/// seam. `main_top` is the already-derived main worktree top ([`linked_worktree_main_root`]).
-/// `Ok(None)` = main has NO `rag-rat.toml` (the local-governs fallback); a main config that
-/// exists but cannot be read or parsed PROPAGATES its error (loading from a linked worktree must
-/// behave like loading from main, errors included). Returns the main worktree TOP as the
-/// governing config dir.
-fn governing_main_config(main_top: &Path) -> Result<Option<(RawConfig, PathBuf)>, ConfigError> {
-    let main_top = main_top.to_path_buf();
-    let main_config_path = main_top.join("rag-rat.toml");
-    let text = match fs::read_to_string(&main_config_path) {
-        Ok(text) => text,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(err) => return Err(err.into()),
-    };
-    let raw: RawConfig = toml::from_str(&text)?;
-    if raw.local_ai.is_some() {
-        return Err(ConfigError::LocalAiTableRenamed);
-    }
-    if raw.dream.is_some() {
-        return Err(ConfigError::DreamTableMoved);
-    }
-    Ok(Some((raw, main_top)))
-}
-
-fn resolve_targets(
-    root: &Path,
-    simple: BTreeMap<String, Vec<String>>,
-    expanded: Vec<RawTarget>,
-) -> Result<Vec<ResolvedTarget>, ConfigError> {
-    let mut names = BTreeSet::new();
-    let mut targets = Vec::new();
-
-    for (language_name, directories) in simple {
-        let language = Language::from_str(&language_name)?;
-        let kind =
-            if language == Language::Markdown { TargetKind::Docs } else { TargetKind::Source };
-        let name = language.as_str().to_string();
-        push_target(root, &mut names, &mut targets, ResolvedTarget {
-            include: language.default_include_globs(),
-            exclude: Vec::new(),
-            name,
-            language,
-            directories: directories.into_iter().map(PathBuf::from).collect(),
-            kind,
-        })?;
-    }
-
-    for target in expanded {
-        let language = Language::from_str(&target.language)?;
-        let kind = target
-            .kind
-            .as_deref()
-            .map(TargetKind::from_str)
-            .transpose()?
-            .unwrap_or(TargetKind::Source);
-        push_target(root, &mut names, &mut targets, ResolvedTarget {
-            name: target.name,
-            language,
-            directories: target.directories.into_iter().map(PathBuf::from).collect(),
-            include: target.include.unwrap_or_else(|| language.default_include_globs()),
-            exclude: target.exclude.unwrap_or_default(),
-            kind,
-        })?;
-    }
-
-    Ok(targets)
-}
-
-fn push_target(
-    root: &Path,
-    names: &mut BTreeSet<String>,
-    targets: &mut Vec<ResolvedTarget>,
-    target: ResolvedTarget,
-) -> Result<(), ConfigError> {
-    if !names.insert(target.name.clone()) {
-        return Err(ConfigError::DuplicateTarget(target.name));
-    }
-    for directory in &target.directories {
-        let full_path = root.join(directory);
-        if !full_path.is_dir() {
-            return Err(ConfigError::MissingDirectory(directory.clone()));
-        }
-    }
-    targets.push(target);
-    Ok(())
-}
-
-/// Re-anchor a **linked** worktree's `root` to the equivalent path under the **main** worktree, so
-/// every worktree of a repo resolves to one root + one shared index — while PRESERVING any
-/// subdirectory the config root points at (a `root="<subdir>"` rebases to `<main>/<subdir>`, not
-/// the repo top). The main worktree (and non-git dirs) resolve to themselves. Collapsing a subdir
-/// root to the repo top changed the indexed file set and could fail config load when a target dir
-/// exists only under the subdir (#219 review).
-fn anchor_root_to_main_worktree(root: &Path) -> PathBuf {
-    let Ok(repo) = crate::index::discover_repo(root) else {
-        return root.to_path_buf();
-    };
-    let (Some(workdir), Some(main_root)) = (repo.workdir(), main_worktree_root(root)) else {
-        return root.to_path_buf();
-    };
-    let workdir = workdir.canonicalize().unwrap_or_else(|_| workdir.to_path_buf());
-    if main_root == workdir {
-        return root.to_path_buf(); // already the main worktree — keep the configured (sub)root
-    }
-    // Linked worktree: rebase root's in-worktree subpath under the main worktree top. `root` is
-    // canonicalized by `normalize_existing_dir`, so it strips cleanly against the canonical
-    // workdir; `root == workdir` (a `root="."` config) yields an empty suffix → the main
-    // worktree top.
-    let anchored = match root.strip_prefix(&workdir) {
-        Ok(rel) => main_root.join(rel),
-        Err(_) => main_root,
-    };
-    // The anchored path must EXIST in the main checkout. When the linked branch sets `[index].root`
-    // to a directory that lives only on the branch (not in main), `main_root.join(rel)` points at a
-    // missing path, which would break the later `discover_repo` / database-base / base-indexing
-    // calls that read `Config.root`. Keep the linked checkout's (validated, existing) root in that
-    // case — the overlay still serves the branch; the base just can't anchor there (#219 review).
-    if anchored.is_dir() { anchored } else { root.to_path_buf() }
-}
+use crate::language::LanguageError;
+
+mod discovery;
+mod load;
+mod raw;
+mod types;
 
 #[derive(Debug, Error)]
 pub enum ConfigError {
@@ -1339,21 +139,29 @@ pub enum ConfigError {
     UnknownMemorySurface(String),
 }
 
-mod discovery;
-mod raw;
-
 pub use discovery::{
     default_database_path, default_legacy_database_path, discover_config_path,
     linked_worktree_main_root, nearest_config_at_or_above,
 };
 pub(crate) use discovery::{main_worktree_root, normalize_existing_dir, resolve_default_database};
+#[cfg(test)]
+pub(crate) use load::{anchor_root_to_main_worktree, resolve_targets};
 pub use raw::endpoint_authority_has_userinfo;
 pub(crate) use raw::{RawConfig, RawTarget, resolve_relative_cookbook_path};
 #[cfg(test)]
 pub(crate) use raw::{RawMemory, RawOracle, RawSearch, RawVersionCheck, RawWatch};
+pub use types::{
+    Config, DEFAULT_QUERY_ENDPOINT, DreamLlmConfig, EmbeddingBackend, EmbeddingConfig,
+    EmbeddingRuntimeConfig, LlmConfig, LogConfig, LogFormat, LogLevel,
+    MAX_REMOTE_EMBEDDING_CONCURRENCY, MemoryConfig, MemorySurface, OracleConfig, RemoteBackend,
+    RemoteDreamConfig, RemoteEmbeddingConfig, ResolvedTarget, SearchConfig, TargetKind,
+    VersionCheckConfig, WatchConfig,
+};
 
 #[cfg(test)]
 mod tests {
+    use std::collections::BTreeMap;
+    use std::path::Path;
     use std::sync::atomic::{AtomicU64, Ordering};
 
     // Test-only: forward-slashes a real filesystem path so it can be embedded in a
@@ -1362,6 +170,7 @@ mod tests {
     use path_slash::PathExt;
 
     use super::*;
+    use crate::language::Language;
 
     static CFG_TEMP: AtomicU64 = AtomicU64::new(0);
